Fix test (#1394)
This pull request refactors how the Apollo Client is instantiated to use
the recommended `HttpLink` instead of passing the `uri` directly. It
also updates the related test to properly mock and verify the
construction of the `HttpLink` and its usage in the `ApolloClient`
setup.

**Apollo Client instantiation improvements:**

* Refactored `src/utils/api/apollo-client.ts` to construct the Apollo
Client using a `HttpLink` with the `uri` set to the GraphQL URL, instead
of passing the `uri` directly to the client.

**Test enhancements:**

* Updated `src/utils/api/apollo-client.test.ts` to mock the `HttpLink`
constructor, verify it is called with the correct URI, and check that
the resulting link instance is passed to the `ApolloClient`. Also
ensured that the cache constructor is still called.